# core/docs/swagger.yaml
definitions:
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  title: AI Proxy Swagger API
  version: "1.0"
paths:
...
  /api/channel/{id}/balance:
    get:
      description: Updates the balance for a single channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  type: number
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update channel balance
      tags:
      - channel
...
  /api/channels/balance:
    get:
      description: Updates the balance for all channels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update all channels balance
      tags:
      - channel
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"